WS07 setRemoteAddr receives u_int32_t

+++ src/CSocket.cpp
@@ -6,7 +6,7 @@
 /*   By: shtanemu <shtanemu@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/09 13:01:41 by shtanemu          #+#    #+#             */
-/*   Updated: 2023/10/17 20:26:29 by shtanemu         ###   ########.fr       */
+/*   Updated: 2023/10/22 19:48:14 by shtanemu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -114,7 +114,7 @@ void CSocket::setLasttime(std::time_t const &_lasttime) {
 	lasttime = _lasttime;
 }
 
-void CSocket::setRemoteAddr(unsigned long s_addr) {
+void CSocket::setRemoteAddr(u_int32_t s_addr) {
 	int byte;
 	int bitshift(32);
 	std::stringstream ss;
